+++ src/commands/contract/compile.ts
@@ -1,9 +1,15 @@
 import { Command, Flags } from "@oclif/core";
-import { spawn } from "node:child_process";
 import path = require("node:path");
 import { readdirSync } from "node:fs";
-import { ensureSwankyProject } from "../../lib/command-utils";
+import {
+  copyArtefactsFor,
+  ensureSwankyProject,
+  getBuildCommandFor,
+  getSwankyConfig,
+} from "../../lib/command-utils";
 import { Spinner } from "../../lib/spinner";
+import { BuildData } from "../../types";
+import { writeJSON } from "fs-extra";
 export class CompileContract extends Command {
   static description = "Compile the smart contract(s) in your contracts directory";
 
@@ -28,34 +34,57 @@ export class CompileContract extends Command {
 
     await ensureSwankyProject();
 
+    const config = await getSwankyConfig();
+
+    const contractInfo = config.contracts[args.contractName];
+    if (!contractInfo) {
+      this.error(`Cannot find a contract named ${args.contractName} in swanky.config.json`);
+    }
+
     const spinner = new Spinner();
-    await new Promise<void>((resolve, reject) => {
-      spinner.start("Compiling contract");
-      const contractList = readdirSync(path.resolve("contracts"));
 
-      if (!contractList.includes(args.contractName)) {
-        throw Error(`Contract name ${args.contractName} is invalid`)
-      }
+    const contractList = readdirSync(path.resolve("contracts"));
 
-      const build = spawn("cargo", ["+nightly", "contract", "build"], {
-        cwd: path.resolve("contracts", args.contractName),
+    const contractPath = path.resolve("contracts", args.contractName);
+    if (!contractList.includes(args.contractName)) {
+      this.error(`Path to contract ${args.contractName} does not exist: ${contractPath}`);
+    }
+
+    await spinner.runCommand(
+      async () => {
+        return new Promise<void>((resolve, reject) => {
+          const build = getBuildCommandFor(contractInfo.language, contractPath);
+
+          build.stdout.on("data", () => spinner.ora.clear());
+          build.stdout.pipe(process.stdout);
+          if (flags.verbose) {
+            build.stderr.on("data", () => spinner.ora.clear());
+            build.stderr.pipe(process.stdout);
+          }
+
+          build.on("error", (error) => {
+            reject(error);
+          });
+
+          build.on("exit", () => {
+            resolve();
+          });
+        });
+      },
+      "Compiling contract",
+      "Contract compiled successfully"
+    );
+
+    const buildData = (await spinner.runCommand(async () => {
+      return copyArtefactsFor(contractInfo.language, contractInfo.name, contractPath);
+    }, "Copying artefacts")) as BuildData;
+
+    await spinner.runCommand(async () => {
+      contractInfo.build = buildData;
+
+      await writeJSON(path.resolve("swanky.config.json"), config, {
+        spaces: 2,
       });
-
-      build.stdout.on("data", () => spinner.ora.clear());
-      build.stdout.pipe(process.stdout);
-      if (flags.verbose) {
-        build.stderr.on("data", () => spinner.ora.clear());
-        build.stderr.pipe(process.stdout);
-      }
-
-      build.on("error", (error) => {
-        reject(error);
-      });
-
-      build.on("exit", () => {
-        resolve();
-      });
-    });
-    spinner.succeed("Contract compiled successfully");
+    }, "Writing config");
   }
 }