+++ src/lib/tasks.ts
@@ -52,15 +52,26 @@ export async function copyContractTemplateFiles(
 
 export async function processTemplates(projectPath: string, templateData: Record<string, string>) {
   const templateFiles = await globby(projectPath, {
-    expandDirectories: { extensions: ["tpl"] },
+    expandDirectories: { extensions: ["hbs"] },
   });
+
+  handlebars.registerHelper("if_eq", function (a, b, options): boolean {
+    if (a === b) {
+      // @ts-ignore
+      return options.fn(this);
+    } else {
+      // @ts-ignore
+      return options.inverse(this);
+    }
+  });
+
   await Promise.all(
     templateFiles.map(async (tplFilePath) => {
       const rawTemplate = await readFile(tplFilePath, "utf8");
       const template = handlebars.compile(rawTemplate);
       const compiledFile = template(templateData);
       await rm(tplFilePath);
-      await writeFile(tplFilePath.split(".tpl")[0], compiledFile);
+      await writeFile(tplFilePath.split(".hbs")[0], compiledFile);
     })
   );
 }