+++ CHANGELOG.md
@@ -1,4 +1,19 @@
-## v0.2.0 (2022-10-19)
+
+## v0.3.0 (2022-11-01)
+
+### Features
+
+- Add Ask! language support and templates
+- Update config schema to properly track builds/deployments
+- store builds in common `/artefacts` folder
+
+### Trivial changes
+
+- [Deployments management](https://github.com/AstarNetwork/swanky-cli/issues/12)
+- [Populate Contracts field on init](https://github.com/AstarNetwork/swanky-cli/issues/21)
+- [Move common methods and types to lib](https://github.com/AstarNetwork/swanky-cli/issues/16)
+
+## v0.2.1 (2022-10-19)
 
 ### Trivial changes
 
@@ -6,8 +21,6 @@
 - update ink! to 3.3.0
 
 ## v0.2.0 (2022-09-22)
-
----
 
 ### Features